거래처 거래 물품 관리 기능 구현

+++ src/main/java/kr/co/easystock/domain/stock/StockRepository.java
@@ -3,6 +3,6 @@ package kr.co.easystock.domain.stock;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface StockRepository extends JpaRepository<Stock, Integer>, JpaSpecificationExecutor<Stock>
+public interface StockRepository extends JpaRepository<Stock, Long>, JpaSpecificationExecutor<Stock>
 {
 }
